Рефакторинг обработки ответа от 1503, изменение на List<Map<>>.

+++ src/main/java/ru/vtb/msa/rfrm/integration/rfrmkafka/model/PayCoreKafkaModel.java
@@ -10,7 +10,7 @@ import lombok.*;
 @AllArgsConstructor
 public class PayCoreKafkaModel {
     @JsonProperty("reward_id")
-    private Integer rewardId;
+    private Long rewardId;
     @JsonProperty("status")
     private Integer status;
     @JsonProperty("status_description")
