+++ src/main/java/ru/vtb/msa/rfrm/processingDatabase/batch/ActionEntPaymentTaskRepo.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface ActionEntPaymentTaskRepo {
 
     // обновление списка rewardId, установка для каждого rewardId blocked=1 и blocked_at=now()
-    void updateBlockByRewardIdEqualOne(List<Integer> entities);
+    void updateBlockByRewardIdEqualOne(List<Long> entities);
 
     // обновление списка rewardId, установка для каждого rewardId blocked=0 и blocked_at=now()
-    void updateBlockByRewardIdEqualZero(List<Integer> entities);
+    void updateBlockByRewardIdEqualZero(List<Long> entities);
 
 }
