+++ src/main/java/ru/vtb/msa/rfrm/service/ProcessClientAccounts.java
@@ -3,5 +3,5 @@ package ru.vtb.msa.rfrm.service;
 import ru.vtb.msa.rfrm.integration.personaccounts.client.model.response.Account;
 
 public interface ProcessClientAccounts {
-    void processAccounts(Account account, String result, Long mdmId, Integer rewardId);
+    void processAccounts(Account account, String result, Long mdmId, Long rewardId);
 }
