+++ src/main/java/ru/vtb/msa/rfrm/service/ServiceAccountsInterface.java
@@ -1,5 +1,5 @@
 package ru.vtb.msa.rfrm.service;
 
 public interface ServiceAccountsInterface {
-    void getClientAccounts(Long mdmIdFromKafka, Integer rewardId);
+    void getClientAccounts(Long mdmIdFromKafka, Long rewardId);
 }
